requirements_cuda10.txt
backcall==0.1.0
chardet==3.0.4
click==7.1.1
cycler==0.10.0
datasketch==1.2.3
decorator==4.4.2
flask==0.12.2
flask-socketio==2.8.5
idna==2.9
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
langdetect==1.0.7
markupsafe==1.1.1
matplotlib==3.5.0
nltk==3.4.5
pandas==0.25.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
pygments==2.6.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.0
sentencepiece==0.1.85
tensorboardx==2.1
torchtext==0.5.0
tqdm==4.43.0
traitlets==4.3.3
ujson==1.35
urllib3==1.25.8
wcwidth==0.1.8
werkzeug==1.0.0
whoosh==2.7.4